buffers the tcp header for reuse, and creates a write method to write to the nic
Rather than creating a new TCP header for every packet recieved, we can use the same header but modify it for every packet we send. We also use a write method that takes a nic and a payload and writes a full ip packet ot the nic - including the payload.