Establishes a connection and initiates its termination
Fixes the 'write' method to calculate the tcp checksum correctly; moved the 'turning on of the ack flag' to the top of the function so it can be calculated in the tcp header checksum, and used the correct payload for checksum calculation.

Modifies the 'on_packet' method to properly handle arriving segments.

Fixes a logical error in the 'is_duplicate' function; if SND.UNA and SND.NXT are equal, then it means that all our sent segments have been ACKed.